src/agents/DigitalTransform/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pyyaml>=6.0.0
requests>=2.31.0
typing-extensions>=4.0.0
colorama>=0.4.6
tqdm>=4.65.0 